requirements.txt
asgiref==3.8.1
Django==5.0.6
django-crispy-forms==2.2
django-extensions==3.2.3
django-filter==24.2
django-widget-tweaks==1.5.0
sqlparse==0.5.0
typing_extensions==4.12.2
